only return updated = true if calls to update are two ms between the update calls
The Joypad driver now reports changes only if either it got
new events, or the buttons which are used to emulate the
slider axis are pressed and the last update call was 2ms ago.

This prevents endless loops if the button to emulate the
slider axis is pressed and the "user" polls until there is
no new update.